requirements.txt
bs4==0.0.2
dataclasses-json==0.6.4
datasets==2.18.0
html2text==2024.2.26
langchain-text-splitters==0.0.1
langchain==0.1.16
loguru==0.7.2
openai==1.23.6
python-dateutil==2.8.2
python-dotenv==1.0.0
ollama==0.3.3
requests-oauthlib==1.3.1
requests==2.32.3
rich==13.9.3
supabase==2.5.0
tiktoken==0.8.0
urllib3==1.26.20
watchtower==3.3.1
weaviate-client==4.6.5
frozendict==2.4.4